+++ v-next.sql
@@ -1,5 +1,7 @@
 \i ocd-2960.sql
 
+\i ocd-2156.sql
+
 \i dev/openchpl_soft-delete.sql
 \i dev/openchpl_views.sql
 \i dev/openchpl_grant-all.sql